+++ xcloud-dao/src/main/java/com/wl4g/devops/dao/ci/PipeHistoryPcmDao.java
@@ -18,13 +18,13 @@ package com.wl4g.devops.dao.ci;
 import com.wl4g.components.core.bean.ci.PipeHistoryPcm;
 
 public interface PipeHistoryPcmDao {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(PipeHistoryPcm record);
 
     int insertSelective(PipeHistoryPcm record);
 
-    PipeHistoryPcm selectByPrimaryKey(Integer id);
+    PipeHistoryPcm selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(PipeHistoryPcm record);
